Update standard.md
Change from: The project must state that it complies with relevant privacy laws, and all applicable international and domestic laws
Change to: The project must state that to the best of its knowledge it complies with relevant privacy laws, and all applicable international and domestic laws.
Reason: It's not clear how DPGs are going to be able to state this with confidence. Referencing the privacy and security standards is one thing (and very possible), but attesting to compliance with domestic laws for all potential implementation countries does not seem possible. 
@eduffus